Extract CLI mock socket test helpers
